+++ exp/services/recoverysigner/internal/crypto/service_key_secure_aws.go
